add same filter to not show group of reports if no reports to show
Conflicts:
	inc/common.class.php

+++ inc/profile.class.php
@@ -266,13 +266,16 @@ class PluginMreportingProfile extends CommonDBTM {
             $mreportingConfig = new PluginMreportingConfig();
             $mreportingConfig->getFromDB($report['id']);
 
-            $profile = $this->findByProfileAndReport($ID,$report['id']);
-            $index = str_replace('PluginMreporting','',$mreportingConfig->fields['classname']);
-            $title = $LANG['plugin_mreporting'][$index][$report['name']]['title'];
+            //If classname doesn't exists, don't display the report
+            if (class_exists($mreportingConfig->fields['classname'])) {
+               $profile = $this->findByProfileAndReport($ID,$report['id']);
+               $index = str_replace('PluginMreporting','',$mreportingConfig->fields['classname']);
+               $title = $LANG['plugin_mreporting'][$index][$report['name']]['title'];
 
-            echo "<tr class='tab_bg_1'><td>" . $mreportingConfig->getLink() . "&nbsp(".$title."): </td><td>";
-            Profile::dropdownNoneReadWrite($report['id'],$profile->fields['right'],1,1,0);
-            echo "</td></tr>\n";
+               echo "<tr class='tab_bg_1'><td>" . $mreportingConfig->getLink() . "&nbsp(".$title."): </td><td>";
+               Profile::dropdownNoneReadWrite($report['id'],$profile->fields['right'],1,1,0);
+               echo "</td></tr>\n";
+            }
         }
